requirements.txt
groq
python-dotenv
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.28.0
bitsandbytes>=0.43.0
peft>=0.10.0
trl>=0.8.0
